Atualização
Modificação da classe de vailidação
